pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["JagTheFriend (jagthefriend12@gmail.com)"]

[tool.poetry.dependencies]
python = "^3.8"
lyrics-extractor = "^3.0.1"
praw = "^7.2.0"
art = "^5.1"
flask = "^1.1.2"
google-api-python-client = "^2.0.2"
wandbox-api = "^0.9.25"
prsaw = "^0.2.0"
python-aiml = "^0.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
